optimize mobile view for featured activity pack view
