+++ services/QuillLMS/client/app/bundles/Teacher/components/assignment_flow/unit_templates_manager/unit_template_profile/__tests__/unit_template_profile_assign_button.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { mount } from 'enzyme';
 
-import { UnitTemplateProfileAssignButton } from '../unit_template_profile_assign_button';
+import { UnitTemplateProfileDisclaimer } from '../unit_template_profile_assign_button';
 import { activityPack, evidenceActivity } from '../../../../../../../test_data/activityPack';
 
 const mockProps = {
@@ -9,7 +9,7 @@ const mockProps = {
 }
 describe('UnitTemplateProfileDescription component', () => {
   let component = mount(
-    <UnitTemplateProfileAssignButton {...mockProps} />
+    <UnitTemplateProfileDisclaimer {...mockProps} />
   );
   it('should match snapshot', ()=> {
     expect(component).toMatchSnapshot();
@@ -17,13 +17,13 @@ describe('UnitTemplateProfileDescription component', () => {
   it('should show login and signup buttons if non_authenticated is true', ()=> {
     expect(component.find('.login-or-signup-buttons').length).toEqual(0);
     mockProps.data.non_authenticated = true;
-    component = mount(<UnitTemplateProfileAssignButton {...mockProps} />);
+    component = mount(<UnitTemplateProfileDisclaimer {...mockProps} />);
     expect(component.find('.login-or-signup-buttons').length).toEqual(1);
   })
   it('should show Evidence warning if at least one Evidence activity is present', ()=> {
     expect(component.find('.evidence-warning').length).toEqual(0);
     mockProps.data.activities.push(evidenceActivity);
-    component = mount(<UnitTemplateProfileAssignButton {...mockProps} />);
+    component = mount(<UnitTemplateProfileDisclaimer {...mockProps} />);
     expect(component.find('.evidence-warning').length).toEqual(1);
   })
 });
